moving method into helper without relying on kafka spring config
